Combine tophat menu components into one
